refactor: extract common global variables to store index
This will reduce resource usage.
For instance, useNow requests a lot of animation frames, which can be heavy when
invoked in multiple components.

Also, removed wildcard exports for stores (types now must be explicitly imported from the relevant store)
This is to avoid polluting the store exports. Just the new globals are allowed for wildcards.